ASoC: fsl_sai: fix u32 overflow in fsl_sai_set_bclk()
freq * 512 overflows u32 when freq > 8,388,607 (e.g. 48kHz/8ch/32bit).
Cast freq to unsigned long to use 64-bit arithmetic on arm64.

error logs:
  fsl-sai 443b0000.sai: failed to derive required Rx rate: 12288000
  fsl-sai 443b0000.sai: ASoC: error at snd_soc_dai_hw_params on sai-tx-rx: -22

+++ sound/soc/fsl/fsl_sai.c
@@ -484,8 +484,8 @@ static int fsl_sai_set_bclk(struct snd_soc_dai *dai, bool tx, u32 freq)
 		 * so we set it to maximum ratio, then scale the frequency
 		 * below 300Mhz, but above stay above lowest ratio
 		 */
-		clk_rate = freq * 512;
-		while (clk_rate >= pll_max_freq*1000*1000 && clk_rate >= freq*2)
+		clk_rate = (unsigned long)freq * 512;
+		while (clk_rate >= pll_max_freq*1000*1000 && clk_rate >= (unsigned long)freq*2)
 			clk_rate /= 2;
 
 		ret = clk_set_rate(sai->mclk_clk[id], clk_rate);
